% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed_versions.R
\name{check_installed_versions}
\alias{check_installed_versions}
\title{check_installed_versions}
\usage{
check_installed_versions(lock_file_loc = "proj_lib.lock")
}
\arguments{
\item{lock_file_loc}{Path to project lock file.}
}
\description{
Validates that the installed package versions match the lock file
}

