% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_profile.R
\name{create_r_profile}
\alias{create_r_profile}
\title{create_r_profile}
\usage{
create_r_profile(mattlib_loc = "mattlib/lib/")
}
\arguments{
\item{mattlib_loc}{Location where mattlib is found}
}
\description{
Creates the project specific Rprofile.
}

