% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_lib.R
\name{deploy_lib}
\alias{deploy_lib}
\title{deploy_lib}
\usage{
deploy_lib(proj_lib = "proj_lib", lock_file_loc = "proj_lib.lock",
  do_thaw = TRUE, redeploy = TRUE)
}
\arguments{
\item{proj_lib}{Directory where src packages should be installed.}

\item{lock_file_loc}{Location where project lockfile is found.}

\item{do_thaw}{Should packmatt download the src files? Default: True}

\item{redeploy}{Should packmatt redploy from scratch? Default: True}
}
\description{
Deploys the package sources to the mattlib and sets an .Rprofile
}

