% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze_packages.R
\name{freeze_packages}
\alias{freeze_packages}
\title{freeze_packages}
\usage{
freeze_packages(lock_file_loc = "mattpack.lock")
}
\arguments{
\item{lock_file_loc}{Location to save the mattpack lock file.}
}
\description{
freeze_packages
}

