% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mattpack_repo.R
\name{install_mattpack_lib}
\alias{install_mattpack_lib}
\title{install_mattpack_lib}
\usage{
install_mattpack_lib(mattlib_loc = "mattlib",
  lock_file_loc = "mattpack.lock", git_pat = NULL)
}
\arguments{
\item{mattlib_loc}{Directory where src packages should be installed.}

\item{lock_file_loc}{Location where mattpack lockfile is found.}

\item{git_pat}{Github Pat.  Optional and only used if downloaded from a private repo.}
}

