% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mattpack_repo.R
\name{install_mattpack_repo}
\alias{install_mattpack_repo}
\title{install_mattpack_repo}
\usage{
install_mattpack_repo(src_download_loc, mattlib_loc, lock_file_loc)
}
\arguments{
\item{src_download_loc}{Location to where package src files are located.}

\item{mattlib_loc}{Directory where src packages should be installed.}

\item{lock_file_loc}{Location where mattpack lockfile is found.}
}

