% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_packages_links.R
\name{prep_package_links}
\alias{prep_package_links}
\title{prep_package_links}
\usage{
prep_package_links(lock_file_loc = "proj_lib.lock")
}
\arguments{
\item{lock_file_loc}{Path to proj_lib lock file.}
}
\description{
Prepares lock file to be collected by generating urls for download.
}

