% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thaw_mattpack.R
\name{thaw_mattpack}
\alias{thaw_mattpack}
\title{thaw_mattpack}
\usage{
thaw_mattpack(lock_file_loc = "mattpack.lock", github_pat = NULL,
  quiet = TRUE)
}
\arguments{
\item{lock_file_loc}{Path to mattpack lock file.}

\item{github_pat}{Github Pat.  Optional and only used if downloaded from a private repo.}

\item{quiet}{Logical; Should download.file return verbose output? Optional.}
}
\description{
Collects source files for all packages specified in lockfile.
}
\details{
thaw_mattpack
}

