% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thaw_proj_lib.R
\name{thaw_proj_lib}
\alias{thaw_proj_lib}
\title{thaw_proj_lib}
\usage{
thaw_proj_lib(lock_file_loc = "proj_lib.lock", github_pat = NA,
  quiet = TRUE)
}
\arguments{
\item{lock_file_loc}{Path to project lock file.}

\item{github_pat}{Github Pat.  Optional and only used if downloaded from a private repo.}

\item{quiet}{Logical; Should download.file return verbose output? Optional.}
}
\description{
Collects source files for all packages specified in lockfile.
}

