% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_r_profile.R
\name{create_r_profile}
\alias{create_r_profile}
\title{create_r_profile}
\usage{
create_r_profile(proj_lib = "proj_lib/lib/")
}
\arguments{
\item{proj_lib}{Location where project library is found}
}
\description{
Creates the project specific Rprofile.
}

