% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_lib.R
\name{deploy_lib}
\alias{deploy_lib}
\title{deploy_lib}
\usage{
deploy_lib(mattlib_loc = "mattlib", lock_file_loc = "mattpack.lock",
  do_thaw = FALSE)
}
\arguments{
\item{mattlib_loc}{Directory where src packages should be installed.}

\item{lock_file_loc}{Location where mattpack lockfile is found.}
}
\description{
Deploys the package sources to the mattlib and sets an .Rprofile
}

