% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_mattpack_repo.R
\name{install_mattpack_repo}
\alias{install_mattpack_repo}
\title{install_mattpack_repo}
\usage{
install_mattpack_repo(mattlib_loc = "mattlib",
  lock_file_lock = "mattpack.lock", git_pat = NULL)
}
\arguments{
\item{mattlib_loc}{Directory where src packages should be installed.}

\item{git_pat}{Github Pat.  Optional and only used if downloaded from a private repo.}

\item{lock_file_loc}{Location where mattpack lockfile is found.}
}

