% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_packages_links.R
\name{prep_package_links}
\alias{prep_package_links}
\title{prep_package_links}
\usage{
prep_package_links(lock_file_loc = "mattpack.lock")
}
\arguments{
\item{lock_file_loc}{Path to mattpack lock file.}
}

