% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thaw_mattpack.R
\name{thaw_mattpack}
\alias{thaw_mattpack}
\title{thaw_mattpack}
\usage{
thaw_mattpack(lock_file_loc, src_download_loc, quiet = TRUE)
}
\arguments{
\item{lock_file_loc}{Path to mattpack lock file.}

\item{src_download_loc}{Directory where package source files should be downloaded.}

\item{quiet}{Logical; Should download.file return verbose output? Optional.}
}
\description{
thaw_mattpack
}

